Add files via upload
Many neighborhoods in Fort George, Inwood, Upper Manhattan, Upper West Side, and Washington Heights have a lot of noise complaints in the north part of Manhattan. Chelsea, West Village, and SoHo areas also show some high volume of noise complaints. There are many reasons towards the noise issue in NYC. As for the Northern Part, the population in Uptown Manhattan is 554,925, whereas the total population of Midtown Manhattan is 104,753 and 298,921 for Downtown Manhattan. As the number shows, the population of Uptown Manhattan is much higher that that of the rest of Manhattan. As a result, the people are more likely to be affected by the loud noise than other areas. The large number of populations also shows higher possibilities for complaints. Another reason is that the large populations are also placed in houses and apartments that are closer to one another, which is also a potential reason for more complaints. Last but not least, the street cultures are also different in Uptown, including music from the cars, which is another source for noise complaints. As for Downtown and Midtown, these are the areas where most of tourists visit, and that can lead to some potential unwanted noise during day times. Additionally, there are many bars and restaurant in these areas, which means that these areas are also louder than residential neighborhoods during both day and night time, especially when the dinning outside program is recently invented after COVID-19.  
	The most common complaints have to be loud noise complaints and Construction related complaints. These types of complaints can be seen through out the city which is easy to predict based on the reasoning provided above. Another common type of complaints is barking or animal related complaints. These complaints are also very common, but they are not as clustered as the other complaints. The reason I believe is that the barking noise might only be heard by certain neighbors and the sounds that animals make does not last very long and people might hear that, but they will probably not call for complaints. Last but not least, Ice Cream Trucks are seen around parks which make sense because parks have a lot of visitors and they all are potential customers. 
	The data I found and used supported my insights in Question 1. The data that I found is related to the buildings in NYC, specifically, Manhattan. I filtered the data to only show buildings that are built before 1980s. The reason I made this filter is that The US Government released the Noise Control Act in 1972s targeting Urban Areas. The basic goal is to create a noise free environment for citizens, and that’s when the citizens started to realize the problem in noise pollutions. The act also led to designers’ more concern with noise problem, which means that the building after that are not as noisy as older buildings. Aside from the loud party and music problem. Constructions are also a major issue with old buildings and that means older buildings might need renovations more often than new buildings and that led to noise complaints related to constructions.